# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: English Grammar Learning API
  version: 0.1.0
  description: Contract-first local API. Feature operations remain contract-only until their roadmap phase.
  license:
    name: Private personal use
    identifier: UNLICENSED
servers:
  - url: http://localhost:3001/api/v1
security:
  - sessionCookie: []
paths:
...
  /auth/login:
    post:
      security: []
      operationId: login
      tags: [identity]
      summary: Log in with the local account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [username, password]
              properties:
                username: { type: string, pattern: '^[a-zA-Z0-9_-]+$', maxLength: 64 }
                password: { type: string, minLength: 8, maxLength: 200 }
      responses:
        '200': { description: Authenticated; sets HttpOnly session cookie }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      operationId: logout
      tags: [identity]
      summary: Invalidate the current session
      responses:
        '204': { description: Session invalidated }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [identity]
      summary: Get the authenticated local user
      responses:
        '200': { description: Current user }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /sessions:
    post:
      operationId: startSession
      tags: [sessions]
      summary: Start a learning session
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartSessionRequest'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionView'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          description: No active curriculum or matching published exercises exist
...
  /curriculum-units/{unitId}/challenge:
    post:
      operationId: startUnitChallenge
      tags: [engagement]
      summary: Start a multi-target challenge for a current-level unit
      parameters:
        - name: unitId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: Challenge and session created or replayed safely
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UnitChallengePlan' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /unit-challenges/{challengeId}:
    get:
      operationId: getUnitChallenge
      tags: [engagement]
      summary: Get owner-only per-target challenge results
      parameters:
        - name: challengeId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Auditable challenge projection
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UnitChallengeView' }
        '404': { $ref: '#/components/responses/NotFound' }
  /unit-challenges/{challengeId}/remediation:
    post:
      operationId: remediateUnitChallenge
      tags: [engagement]
      summary: Start focused practice for challenge targets needing practice
      parameters:
        - name: challengeId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: Focused remediation session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionView' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /me/story:
    get:
      operationId: getStoryJourney
      tags: [engagement]
      summary: Get or start the owner A1 story journey
      responses:
        '200':
          description: Current versioned scene and bounded story memory
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StoryJourneyView' }
        '404': { $ref: '#/components/responses/NotFound' }
  /me/story/scenes/{sceneId}/choices:
    post:
      operationId: chooseStoryBranch
      tags: [engagement]
      summary: Record one allowed branch choice idempotently
      parameters:
        - name: sceneId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [choiceId]
              properties:
                choiceId: { type: string, format: uuid }
      responses:
        '201':
          description: Updated story journey
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StoryJourneyView' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /me/story/scenes/{sceneId}/continue:
    post:
      operationId: continueStoryScene
      tags: [engagement]
      summary: Continue or finish a linear scene idempotently
      parameters:
        - name: sceneId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: Updated story journey
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StoryJourneyView' }
        '409': { $ref: '#/components/responses/Conflict' }
  /me/story/scenes/{sceneId}/practice:
    post:
      operationId: practiceStoryScene
      tags: [engagement]
      summary: Start the exact published exercise pinned to the current scene
      parameters:
        - name: sceneId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: Story-linked focused session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionView' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /session-items/{itemId}/attempts:
    post:
      operationId: submitAttempt
      tags: [attempts]
      summary: Submit a new immutable attempt
      parameters:
        - $ref: '#/components/parameters/ItemId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitAttemptRequest'
      responses:
        '202':
          description: Attempt accepted for asynchronous evaluation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttemptAccepted'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /attempts/{attemptId}:
    get:
      operationId: getAttempt
      tags: [attempts]
      summary: Read an owned attempt and its effective evaluation
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Attempt state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttemptView'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: english_session
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    ItemId:
      name: itemId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Authentication is required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    Conflict:
      description: Conflicting state or reused idempotency key with a different payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource was not found or is not owned by the current user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    StartSessionRequest:
      type: object
      additionalProperties: false
      required: [mode]
      properties:
        mode: { type: string, enum: [DAILY, FOCUSED, REVIEW] }
        grammarPointIds:
          type: array
          maxItems: 20
          items: { type: string, format: uuid }
        targetMinutes: { type: integer, minimum: 2, maximum: 30, default: 10 }
...
    UnitChallengePlan:
      type: object
      additionalProperties: false
      required: [challengeId, sessionId]
      properties:
        challengeId: { type: string, format: uuid }
        sessionId: { type: string, format: uuid }
    UnitChallengeView:
      type: object
      additionalProperties: false
      required:
        [
          id,
          unitId,
          unitTitle,
          sessionId,
          status,
          policyVersion,
          startedAt,
          completedAt,
          targets,
          remediationGrammarPointIds,
        ]
      properties:
        id: { type: string, format: uuid }
        unitId: { type: string, format: uuid }
        unitTitle: { type: string }
        sessionId: { type: string, format: uuid }
        status: { type: string, enum: [ACTIVE, COMPLETED] }
        policyVersion: { type: string, const: unit-challenge-v1 }
        startedAt: { type: string, format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        targets:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              [
                grammarPointId,
                grammarCode,
                grammarTitle,
                outcome,
                disposition,
                attemptId,
                reasonCodes,
              ]
            properties:
              grammarPointId: { type: string, format: uuid }
              grammarCode: { type: string }
              grammarTitle: { type: string }
              outcome: { type: string, enum: [PASSED, NEEDS_PRACTICE, NO_EVIDENCE] }
              disposition: { type: [string, 'null'] }
              attemptId: { type: [string, 'null'], format: uuid }
              reasonCodes: { type: array, items: { type: string } }
        remediationGrammarPointIds:
          type: array
          uniqueItems: true
          items: { type: string, format: uuid }
    StoryJourneyView:
      type: object
      additionalProperties: false
      required: [series, status, completedSceneCount, totalSceneCount, currentScene, memoryFacts]
      properties:
        series:
          type: object
          additionalProperties: false
          required: [id, code, title, description, cefr]
          properties:
            id: { type: string, format: uuid }
            code: { type: string }
            title: { type: string }
            description: { type: string }
            cefr: { type: string, enum: [A1, A2, B1, B2, C1, C2] }
        status: { type: string, enum: [ACTIVE, COMPLETED] }
        completedSceneCount: { type: integer, minimum: 0 }
        totalSceneCount: { type: integer, minimum: 1 }
        currentScene: { type: [object, 'null'] }
        memoryFacts:
          type: array
          maxItems: 20
          items:
            type: object
            additionalProperties: false
            required: [key, value]
            properties:
              key: { type: string, maxLength: 64 }
              value: { type: string, maxLength: 160 }
...
    SessionView:
      type: object
      additionalProperties: false
      required: [id, status, mode, startedAt]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [ACTIVE, COMPLETED, ABANDONED] }
        mode: { type: string, enum: [DAILY, FOCUSED, REVIEW] }
        startedAt: { type: string, format: date-time }
...
    SubmitAttemptRequest:
      type: object
      additionalProperties: false
      required: [answer]
      properties:
        answer: { type: string, minLength: 1, maxLength: 2000 }
        clientSubmittedAt: { type: string, format: date-time }
    AttemptAccepted:
      type: object
      additionalProperties: false
      required: [attemptId, status]
      properties:
        attemptId: { type: string, format: uuid }
        status:
          type: string
          enum: [SUBMITTED, EVALUATING, EVALUATED, NEEDS_REVIEW, FAILED]
        evaluation: { $ref: '#/components/schemas/EvaluationView' }
        pollAfterMs: { type: integer, minimum: 250, maximum: 10000 }
    AttemptView:
      type: object
      additionalProperties: false
      required: [attemptId, status]
      properties:
        attemptId: { type: string, format: uuid }
        status:
          type: string
          enum: [SUBMITTED, EVALUATING, EVALUATED, NEEDS_REVIEW, FAILED]
        evaluation:
          $ref: '#/components/schemas/EvaluationView'
...
    EvaluationView:
      type: object
      additionalProperties: false
      required: [disposition, dimensions, feedbackVi, findings, canRetry]
      properties:
        disposition:
          type: string
          enum: [ACCEPT, ACCEPT_WITH_FEEDBACK, RETRY, SYSTEM_REVIEW]
        dimensions:
          type: object
          additionalProperties: false
          required: [targetGrammar, meaning, otherGrammar]
          properties:
            targetGrammar: { $ref: '#/components/schemas/DimensionStatus' }
            meaning: { $ref: '#/components/schemas/DimensionStatus' }
            otherGrammar: { $ref: '#/components/schemas/DimensionStatus' }
        feedbackVi: { type: string }
        findings:
          type: array
          items: { $ref: '#/components/schemas/EvaluationFinding' }
        correctedAnswer: { type: string }
        canRetry: { type: boolean }
    DimensionStatus:
      type: string
      enum: [PASS, MINOR_ISSUES, FAIL, UNCERTAIN]
    EvaluationFinding:
      type: object
      additionalProperties: false
      required: [category, severity, messageVi]
      properties:
        category:
          type: string
          enum: [TARGET_GRAMMAR, OTHER_GRAMMAR, MEANING, VOCABULARY, SPELLING, PUNCTUATION, STYLE]
        severity: { type: string, enum: [INFO, MINOR, MAJOR, BLOCKING] }
        messageVi: { type: string }
        suggestedFix: { type: string }
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code: { type: string, pattern: '^[A-Z][A-Z0-9_]+$' }
            message: { type: string }
            requestId: { type: string, format: uuid }
            details:
              type: array
              items:
                type: object
                additionalProperties: false
                required: [reason]
                properties:
                  field: { type: string }
                  reason: { type: string }